Let RpcCoreService be the AsyncService
